requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
djangorestframework==3.12.4
idna==3.2
iniconfig==1.1.1
jsonschema==3.2.0
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
dj-database-url==0.5.0
python-decouple==3.5
